Show LV2 class in webconf view, and make filterable
This commit also adds incremental search to the filter bar, so it's even
easier to only show one category.

The CSS class of each row in the view is set back to the normal
bootstrap "row", so that if cells overflow, the layout doesn't
break. Without it, particularly long texts in the "type" column will
break the table.