View contents of ZS3 presets
This commit changes the ZS3 section of the "Info" popup for a snapshot
from just listing which presets exist, to showing the exact values they
contain in a table.

Additionally, values that are the same as the base snapshot controller
values are hidden, since those most probably have not been touched by
the user.