Add Tascam US-2x2HR USB interface
Added with a samplerate of 48000. But out of the box the unit is set
to have a samplerate of 44100.

One can change the samplerate of the interface via a Windows machine
with mmsys.cpl to 48000 (it supports samplerates up to 192000).

Or use -r 44100 instead. I do not know whether it is possible to
change its settings via Linux, but changing the rate to something else
in jackd just means you'll get a non-working system.